+++ serialization/go/json/json_parse_node.go
@@ -100,6 +100,16 @@ func loadJsonTree(decoder *json.Decoder) (*JsonParseNode, error) {
 			v := token.(bool)
 			c.SetValue(&v)
 			return c, nil
+		case int8:
+			c := &JsonParseNode{}
+			v := token.(int8)
+			c.SetValue(&v)
+			return c, nil
+		case byte:
+			c := &JsonParseNode{}
+			v := token.(byte)
+			c.SetValue(&v)
+			return c, nil
 		case float64:
 			c := &JsonParseNode{}
 			v := token.(float64)
@@ -243,6 +253,10 @@ func (n *JsonParseNode) getPrimitiveValue(targetType string) (interface{}, error
 		return n.GetStringValue()
 	case "bool":
 		return n.GetBoolValue()
+	case "uint8":
+		return n.GetInt8Value()
+	case "byte":
+		return n.GetByteValue()
 	case "float32":
 		return n.GetFloat32Value()
 	case "float64":
@@ -305,6 +319,22 @@ func (n *JsonParseNode) GetBoolValue() (*bool, error) {
 		return nil, nil
 	}
 	return n.value.(*bool), nil
+}
+
+// GetInt8Value returns a int8 value from the nodes.
+func (n *JsonParseNode) GetInt8Value() (*int8, error) {
+	if n == nil || n.value == nil {
+		return nil, nil
+	}
+	return n.value.(*int8), nil
+}
+
+// GetBoolValue returns a Bool value from the nodes.
+func (n *JsonParseNode) GetByteValue() (*byte, error) {
+	if n == nil || n.value == nil {
+		return nil, nil
+	}
+	return n.value.(*byte), nil
 }
 
 // GetFloat32Value returns a Float32 value from the nodes.