+++ serialization/java/json/lib/src/main/java/com/microsoft/kiota/serialization/JsonSerializationWriter.java
@@ -6,6 +6,7 @@ import com.microsoft.kiota.serialization.Parsable;
 
 import java.lang.Enum;
 import java.lang.reflect.Field;
+import java.math.BigDecimal;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -16,15 +17,11 @@ import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.time.Period;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Base64;
 import java.util.EnumSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.BiConsumer;
 
@@ -51,6 +48,39 @@ public class JsonSerializationWriter implements SerializationWriter {
             }
     }
     public void writeBooleanValue(final String key, final Boolean value) {
+        if(value != null)
+            try {
+                if(key != null && !key.isEmpty()) {
+                    writer.name(key);
+                }
+                writer.value(value);
+            } catch (IOException ex) {
+                throw new RuntimeException("could not serialize value", ex);
+            }
+    }
+    public void writeShortValue(final String key, final Short value) {
+        if(value != null)
+            try {
+                if(key != null && !key.isEmpty()) {
+                    writer.name(key);
+                }
+                writer.value(value);
+            } catch (IOException ex) {
+                throw new RuntimeException("could not serialize value", ex);
+            }
+    }
+    public void writeByteValue(final String key, final Byte value) {
+        if(value != null)
+            try {
+                if(key != null && !key.isEmpty()) {
+                    writer.name(key);
+                }
+                writer.value(value);
+            } catch (IOException ex) {
+                throw new RuntimeException("could not serialize value", ex);
+            }
+    }
+    public void writeBigDecimalValue(final String key, final BigDecimal value) {
         if(value != null)
             try {
                 if(key != null && !key.isEmpty()) {
@@ -303,6 +333,12 @@ public class JsonSerializationWriter implements SerializationWriter {
                 this.writeStringValue(key, (String)value);
             else if(valueClass.equals(Boolean.class))
                 this.writeBooleanValue(key, (Boolean)value);
+            else if(valueClass.equals(Byte.class))
+                this.writeByteValue(key, (Byte)value);
+            else if(valueClass.equals(Short.class))
+                this.writeShortValue(key, (Short)value);
+            else if(valueClass.equals(BigDecimal.class))
+                this.writeBigDecimalValue(key, (BigDecimal)value);
             else if(valueClass.equals(Float.class))
                 this.writeFloatValue(key, (Float)value);
             else if(valueClass.equals(Long.class))